// src/app/modules/admission/models/fee-bill-generation-process.ts
export class StudentListParams {
    ORG_CODE: string;
    CAMPUS_CODE: string;
    USER_CODE : string;
    CLASS_CODE: string;
    VERSION_CODE: string;
    GROUP_CODE: string;
    SESSION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    SHIFT_CODE: string;
    SECTION_CODE: string;
}